requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
ruff==0.6.9
openai==1.57.0
